player controller

+++ src/Constants.java
@@ -5,10 +5,10 @@ public class Constants {
   public static final int SCREEN_HEIGHT = 600;
   public static final String SCREEN_TITLE = "pong";
 
-  public static final int PADDLE_WIDTH = 20;
-  public static final int PADDLE_HEIGHT = 100;
+  public static final double PADDLE_WIDTH = 20;
+  public static final double PADDLE_HEIGHT = 100;
   public static final Color PADDLE_COLOR = Color.WHITE;
-  public static final int PADDING_X = 40;
+  public static final double PADDING_X = 40;
 
-  public static final int BALL_WIDTH = 20;
+  public static final double BALL_WIDTH = 20;
 }
